requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
matplotlib>=3.4.0
opencv-python>=4.5.0
Pillow>=8.0.0
pytest>=7.0.0
